LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/mixin-media-query-invalidation-2-expected.txt



FAIL CSS Mixins: Invalidates on negative media query becoming positive assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 0, 0)"
